[Y26W2-305] refactor(web): Remove unused style props, refactor to use index.tsx
