+++ apps/web/src/domains/landing/components/gallery-section/index.tsx
