+++ apps/web/src/domains/landing/components/problem-banner-section/index.tsx
@@ -1,4 +1,4 @@
-import LandingSectionContainer from "./landing-section-container";
+import LandingSectionContainer from "../landing-section-container";
 
 const ProblemBannerSection = () => {
   return (
